+++ cpp/src/io/orc/writer_impl.cu
@@ -1636,13 +1636,13 @@ encoder_decimal_info decimal_chunk_sizes(orc_table_view& orc_table,
                          if (col.is_null(idx) or not bit_value_or(pushdown_mask, idx, true))
                            return 0u;
 
-                         int64_t const element =
+                         __int128_t const element =
                            col.type().id() == type_id::DECIMAL32   ? col.element<int32_t>(idx)
                            : col.type().id() == type_id::DECIMAL64 ? col.element<int64_t>(idx)
                                                                    : col.element<__int128_t>(idx);
 
-                         int64_t const sign      = (element < 0) ? 1 : 0;
-                         uint64_t zigzaged_value = ((element ^ -sign) * 2) + sign;
+                         __int128_t const sign      = (element < 0) ? 1 : 0;
+                         __uint128_t zigzaged_value = ((element ^ -sign) * 2) + sign;
 
                          uint32_t encoded_length = 1;
                          while (zigzaged_value > 127) {
